pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django>=5.2.1",
    "django-crispy-forms>=2.4",
    "pillow>=11.2.1",
]
